pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-speaker-diarization-with-adaptive-boundary-refinement"
version = "0.1.0"
description = "Multi-speaker diarization system with adaptive boundary refinement using local acoustic consistency scoring"
authors = [{name = "Alireza Shojaei"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyannote.audio>=3.1.0",
    "pyannote.core>=5.0.0",
    "pyannote.metrics>=3.2.0",
    "speechbrain>=0.5.0",
    "transformers>=4.30.0",
    "pyyaml>=6.0",
    "mlflow>=2.8.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
